+++ setup_wonambi.py
@@ -231,18 +231,19 @@ def _get_files():
 
             if remote['zipped'] is None:
                 print('Copying ' + str(temp_file) + ' to ' + str(final_file))
-                copyfile(temp_file, final_file)  # or maybe symlink
+                copyfile(str(temp_file), str(final_file))  # or maybe symlink
 
             elif remote['zipped'] is True:  # explicit testing
                 with ZipFile(str(temp_file)) as zf:
                     print('Extracting all files in ' + remote['cached'] + ':\n\t' + '\n\t'.join(zf.namelist()))
-                    zf.extractall(path=DATA_PATH)
+                    zf.extractall(path=str(DATA_PATH))
 
             else:
                 print('Extracting file ' + remote['zipped'] + ' to ' +
                       str(final_file))
                 with ZipFile(str(temp_file)) as zf:
-                    extracted = Path(zf.extract(remote['zipped'], path=DOWNLOADS_PATH))
+                    extracted = Path(
+                        zf.extract(remote['zipped'], path=str(DOWNLOADS_PATH)))
                     extracted.rename(final_file)
 
     return 0