+++ wonambi/widgets/utils.py
@@ -30,7 +30,8 @@ stdicon = QCommonStyle.standardIcon
 icon_path = join(dirname(realpath(__file__)), 'icons')
 oxy_path = join(icon_path, 'oxygen')
 
-ICON = {'open_rec': join(oxy_path, 'document-open.png'),
+ICON = {'application': join(icon_path, 'wonambi.jpg'),
+        'open_rec': join(oxy_path, 'document-open.png'),
         'page_prev': join(oxy_path, 'go-previous-view.png'),
         'page_next': join(oxy_path, 'go-next-view.png'),
         'step_prev': join(oxy_path, 'go-previous.png'),
@@ -77,7 +78,7 @@ class Path(QPainterPath):
 
 class RectMarker(QGraphicsRectItem):
     """Class to draw a rectangular, coloured item.
-    
+
     Parameters
     ----------
     x : float
@@ -93,7 +94,7 @@ class RectMarker(QGraphicsRectItem):
     """
     def __init__(self, x, y, width, height, zvalue, color='blue'):
         super().__init__()
-        
+
         self.color = color
         self.setZValue(zvalue)
         buffer = 1
@@ -103,7 +104,7 @@ class RectMarker(QGraphicsRectItem):
 
     def boundingRect(self):
         return self.b_rect
-        
+
     def paint(self, painter, option, widget):
         color = QColor(self.color)
         painter.setBrush(QBrush(color))
@@ -168,7 +169,7 @@ def keep_recent_datasets(max_dataset_history, info=None):
             lg.debug('Removing last dataset ' + history[-1])
             history.pop()
 
-        lg.info('Adding ' + new_dataset + ' to list of recent datasets')
+        lg.debug('Adding ' + new_dataset + ' to list of recent datasets')
         history.insert(0, new_dataset)
         settings.setValue('recent_recordings', history)
         return None