+++ CHANGES.rst
@@ -1,5 +1,12 @@
 Change Log
 ==========
+Version 4
+----------
+- **4.04**: use variable epoch length for scoring, better handling of ABF2 files
+- **4.03**: change nan to zero so we can at least plot something if there are missing data 
+- **4.02**: you can open a dataset from the command line (and better logging)
+- **4.01**: frequency analysis is much more consistent and correct
+
 Version 3
 ----------
 - **3.11**: fixed event marking; stage now visible when zoomed in <30s